make OSGB36 bounding-box bigger
Make the range of allowed latitudes and longitudes bigger.

This is to make it possible to convert all coordinates in
the Ordnance Survey's OSTN15/OSGM15 test set.

The test set is available here:
https://www.ordnancesurvey.co.uk/docs/gps/OSTN15-OSGM15_DevelopersPack.zip

In particular points TP01, TP31 and TP32 where outside of the
bounding-box.

The points are:

PointID ETRS89 Latitude  ETRS Longitude

TP01    49.92226393730  -6.29977752014,100.000
TP31    57.81351838410  -8.57854456076,100.001
TP32    58.21262247180  -7.59255560556,140.404

(from OSTN15_OSGM15_TestInput_ETRStoOSGB.txt file in the zip file above)

so that all coordinates from this